cut!: remove managers and supporting database classes
The manager classes tend to not be the most descriptive way to structure
code.  While they do exist in spirit (see Stager class) having classes
that are not just simple managers allows for better code layout and
readability.  While the code in the managers has yet to be replicated,
removing them now will allow a testable product to be produced sooner.

In addition, the database classes which existed to support the manager
classes are being removed since they no longer represent how the app
will be interacting with the database.