refactor!: change databases from virtual classes to concepts
The databases were created as virtual classes to allow them to be
swapped out for testing and upgrading purposes.  There is no reason to
switch implementations during runtime, so it is a better option to have
them be concepts and remove the virtual class overhead.

This requires making the stager a templated class, so the implementation
 was moved from a .cpp file to a .tpp file and included in the header.