fix(database:staged): change staged directory id name
The staged directory table is the only table which uses a name other
than 'id', rename 'directory_id' to 'id' to bring it in line with the
other tables. There was an error in the primary key for the table caused
by this discrepancy.

Also reformat the archiver_database.sql file to conform to current code
formatting settings.