refactor(database)!: restructure database into BCNF
Having the database in BCNF form is more flexible than the prior
structure. This change does require a change to the StagedDatabase
implementation since some of the data is now in different tables.

Several other changes were performed which accompany this change.
- The hashes and size fields of file_revision can now be null to allow
duplicate revisions to be listed in the file_revision table without
re-specifying the hashes and size.

- The file_duplicate table is now the file_revision_duplicate table and
is simply a joining table. This simplifies finding duplicates since the
file_revision and file_duplicate no long have to be sorted based on the
archive time.

- Typos involving AUTO_INCREMENT were fixed.