build(test): generate random test files
Since storing test files in git would take a lot of space and there are
no dependencies on their contents, generate them from random data each
during project configuration. The size and hash of the random files
needs to be known so get that data and store it in a header file.

This will currently only work on linux based machines and there is
currently no test to ensure that the required commands are available.