refactor: remove redundant constructors and functions
StagedDirectory, StagedFile, and ArchivedDirectory are all structures
yet have a constructor which only initializes the data members and both
StagedDirectory and ArchivedDirectory have getters for these member
functions. These constructors and functions are redundant and should be
removed. In order to reduce the impact of changes the data member of
these structures were rearranged to match the order of the constructors.