fix(config): wrap type errors when loading configuration
If an incorrect type is put into the config file then an exception is
thrown when that config file is loaded, but the default exception
doesn't have any specific information. Wrap calls to retrieve the values
from the json object to capture this exception and throw a more
appropriate exception.