fix(database): the id of the root directory can't be guaranteed
While the name of the root directory will always be '/' it is not a good
idea to rely on the id (which is auto generated by the database) to
always match what the program expects.

For the staged directory database the database isn't guaranteed to have
the root directory since it isn't added by the sql file that is used to
make the database, and it is removed when removing all staged
directories. While it is possible to make sure that the root directory
is always in the staged directory database but this means that when no
directories have been staged the staged directory database still has one
staged directory. This case would be confusing to work with and maintain
and as such the alternative was chosen.