fix(test): incorrect test assumption
When adding files to the archived file database during testing the files
are likely to be added within the same minute resulting in the stored
time being the same. However, this can not be guaranteed since it could
be the case that adding the file the first time happens at the end of
on minute while the second adding of the file happens at the start of
the next minute. Since it can not be guaranteed the test should be
removed.