fix(mock-database): adding archived file gives wrong return value
The file vector iterator is not valid at the end of the function since
the file vector could have has the new file appended which would
invalidate the iterator. Further the return value is being determined by
if the file is already archived when it should be determined by if there
is a file revision with the same hash and size.