fix: archiver and stager using hashes instead of ids
The archiver and stager both use the hash to name the files in the
filesystem, however this can make the filename longer than what the
filesystem supports. This results in hard to diagnose errors, which is
why the tests for the archiver and stager specify that an id value
should be used.

For the stager:
 -The filename of the staged file should be its staged id.
 -When adding a file to the staged directory the staged file should be
 returned.

For the archiver:
 -The filename of the archived file should be the corresponding revision
  id.
 -When adding a file to the archived directory the revision id should be
  returned along with the added type.