fix(mysql-database): destructors are implicitly noexcept
One possible fix is to mark the destructors as noexcept(false) however
this requires additional checks to make sure that no exception was
already thrown as throwing a second exception will terminate the
program.

Instead, if there is an error rolling back the database from the
destructor simply log the error and continue. If the destructor is
running then rolling back the database is not strictly needed as when
the database connection is destroyed the transaction will be abandoned,
but it is not a good idea to rely on this behavior.