test: add tests for RawFile
Also change the requirements in the database tests to check against the
raw file instead of against the compiled constants for hash and size.
This is to have better isolation since those requirements should be that
the database stored the values correctly not that they were calculated
correctly. This isn't (easily) possible for the stager and archiver
tests since those archive the whole directory and do not construct
individual RawFiles at the test case scope.