refactor: merge file, directory, and archive databases
The implementation of the MySQL and mock databases do not separate file,
directory, or archive databases for either of the staged or archived
databases. The primary reason for this is that there is a fair amount
of interaction that needs to happen between the different database that
is not exposed via the api for the databases. While it would be possible
in theory to split them apart, in practice it is not worth the added
hassle and would only serve to make the application more complex without
any benefit. Merging the directory concepts to fall in line with the
implementations will make the code easier to maintain and remove
complexity from database transaction and the construction of stagers and
archivers.