chore: remove unused and unneeded code and modules
There are several functions and code fragments which exist in the code
base but are not used and are not planned to be used within the project,
these should be removed.

There are a couple of functions which are simply wrappers for functions
in the standard library (specifically the isFile, isDirectory, and
pathExists functions). These functions don't add any value and could
cause issues in the future. There are places where the std function is
called directly in code instead of the wrapper version, so if the
wrapper were to change its internal details (the exact std function it
wraps) subtle bugs could be introduced into the project. To prevent this
remove the wrappers and use the std functions directly.

The unfmt library was included in this project because the original
design required the filename of archived file revisions to include the
file id and time stamp and those values would have to be parsed from the
filename string. The current design has no such requirement and as such
has no use for the unfmt library.