refactor(test): restructure database tests to be more clear
Separate the staged database and archived database tests which, while
adding a small amount of duplicated code, drastically increases the
readability of the tests and allows the section names to be more
descriptive.

Commit 12a4550 merged the file, directory, and archive databases, the
tests need to be updated to match that change.

Checking if all the databases are empty is a prerequisite for most tests
so those checks into a single shared function to increase readability.

Change how database tests are included in the build by no longer
excluding them if no database username/password was provided. This is to
allow the mock databases be tested even if the MySQL databases can't be.
If not database/username is provided the build will give a warning but
will continue and at test time the tests for MySQL databases will fail
on connection.