fix(compressor): files not being added to compressed archives
When compressing the archives the compression command is not being given
the correct path to the archive directory. The path should be the
archive path with the archive id added on, but currently only the
archive id is being given which results in the directory not being added
to the archive.