fix(dearchiver): archives are decompressed multiple times
After an archive is decompressed it is supposed to be added to a list so
that the same archive is not decompressed multiple times. However, the
decompressed archives are not being added to that list resulting in
every file causing a decompression. Adding the archive the list solves
this problem.