add!: archive operation
Adding a way to record which file revisions and directories were added
at the same time will be useful for finding/dearchiving specific
revisions for related files. This is currently not possible since there
is no grouping and there is no way to tell which files where added in
which running of the archiver by the revision time.

Adding the archive operation fulfills this need, and also the archive
time to be abstracted out of the file revisions.