refactor!: use inheritance instead of templates for databases
Using templates for the databases is not appropriate as the databases
are required to have the same interface and share a significant amount
of code. By using inheritance this common code can be factored out
making maintaining the database code easier.

To facilitate this transition two new macros 'abstract' and 'interface'
are created to make the code read better.