refactor(getFileReadBuffer): take a vector of sizes instead of config
Passing the entire config object when only a single value is needed
makes the code harder to read and reason about, and harder to test.
Changing the function to take the list of sizes directly means that it
is clear what configuration the function is using, and means that a
config doesn't have to be constructed for testing the function.