fix: empty vector breaks formatting
Passing an empty vector to be formatted causes a segmentation fault as
the formatting code does not test if the vector is empty before
dereferencing the iterator.

While fixing this issue also surround the formatted vector with currly
brackets and add a space after the seperator to make the output more
readable.