test: change how RawFile tests work
When testing that the RawFile objects are created correctly we don't
actually care about the specific hash value of a given file. Instead it
only matters that the hash is unique, not empty, and that the same file
(or a file with the same content) produces the same hash. By removing
the reliance on precomputed hashs in the test, the test later can be
expanded to run on more files with minimal changes.

Similarly c++ has a function to get the size of a file, so use that
instead of a precomputed file size.