fix(archive-directory): archived directories need an archive operation
When a file is archived it has an associated archive operation, the same
goes for directories when they are first archived.  However, when the
same directory is archived again it does not have an associated archive
operation, which means that all directories must be dearchived even if
only one is needed.