add: check if an archive has already been decompressed
Add a check to see if the file revision we are dearchiving has already
been decompressed by a previous execution. This allows the dearchive
command to be called multiple times without having to decompress
archives multiple times as long as the required file revisions have
bene decompressed. Since archived revisions are immutable then even
if the archive may have had additional files added to it, if the file
revision is found then that revision must be valid.