feat: add a check command
The check command should check all archived file revisions against
their respective hashes to ensure that they have been archived and
decompressed without error.

Whether or not an archived file revision is a duplicate or not must
be added to ArchivedFileRevision to allow the check command to skip
revisions which are duplicates of other revisions.