fix(dearchiver): remove duplicate directories during dearchiving
When no archive operation is specified the child directory list
contains multiple copies of archived directories when a child
directory has been archived multiple times. These duplicates must be
removed in order to prevent an attempt to dearchive the same files
multiple times.

Additionally add a check to prevent visiting the root directory as a
child.