fix: have reconstituted archives be stored in the temp directory
When the parts of an archive are put together into a single file that
file should be stored in the temp archive directory instead of the
primary archive directory.

The compressor must take a vector of possible archive locations to
allow it to find archives when decompressing, since single file
archives are not reconstituted and therefore not in the temp archive
directory.