fix(test): stateful lambdas cause errors with some compilers
A stateful lambda is allowed by the C++ standard, and using the
capture list to create local variables for the lambda is also allowed.
However, some compilers generate warnings when this is done, which
with our build environment are treated as errors causing the build to
fail. To fix this the lambda is replaced with a functor.