fix(mock-database): filename without extension gets the wrong archive
If a filename doesn't have an extension or its last character is '.',
then the blank archive should be retrieved. However, currently, the mock
database tries to find an archive for the '' extension, which doesn't
exist. To fix this we check if the extension is blank and manually set
the extension we are looking for to "<BLANK>".